% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsshinyextras-package.R
\docType{package}
\name{phsshinyextras-package}
\alias{phsshinyextras}
\alias{phsshinyextras-package}
\title{phsshinyextras: Function bundle for PHS Shiny Dashboards}
\description{
This package contains multiple functions. There's a range of generic chart functions to keep formatting consistent. There's a generic, downloadable data table. There are various data processing functions too (removing SPSS labels, applying suppression).
}
\author{
\strong{Maintainer}: Ruben Vine \email{ruben.vine@phs.scot}

}
\keyword{internal}
