% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_suppression.R,
%   R/secondary_suppression.R
\name{primary_suppression}
\alias{primary_suppression}
\alias{secondary_suppression}
\title{Apply primary and secondary suppression}
\usage{
primary_suppression(
  data,
  suppression_variable,
  no_secondary = FALSE,
  min_unsuppressed = 5,
  keep_original = TRUE
)

secondary_suppression(data, suppression_variable, ...)
}
\arguments{
\item{data}{A dataset to suppress}

\item{suppression_variable}{A numeric/integer column to suppress}

\item{no_secondary}{TRUE means you don't need to apply secondary suppression}

\item{min_unsuppressed}{The minimum number that should not be suppressed
5 as a default (which means numbers between 1-4 will be suppressed)}

\item{keep_original}{TRUE or FALSE whether you would like a version of the
original column to check suppression has been applied correctly}

\item{...}{A list of variables to group the data by before suppressing
(see examples for usage)}
}
\description{
\code{primary_suppression()} and \code{secondary_suppression()} suppress values in
a column based on disclosure control principles.
}
\details{
\code{primary_suppression()} will apply primary suppression to a numeric or integer
column.
It is designed to be used in tandem with \code{secondary_suppression()} by default.
It converts values ready to be suppressed to -999 temporarily, which
get converted to an asterisk in the \code{secondary_suppression()} function.
To only use primary suppression, use \code{no_secondary = TRUE} as an input.
This will convert suppressed values to an asterisk in the first instance.

By default, the original column is retained for QA purposes and suppression
is applied to numbers between 1-4. There are inputs to change these defaults
\code{secondary_suppression()} is applied after primary suppression to data that
displays a total (such as having SIMD 1-5 and Total).

Secondary suppression ensures a suppressed value cannot be calculated when
only one value is suppressed. It does so by suppressing the next lowest value
when only one number is suppressed.
}
\examples{

# Load the sample_data
data("sample_data")

# Apply primary suppression across the board, with no secondary suppression
suppresssed_data1 = sample_data |>
  primary_suppression(suppression_variable = number,
                      no_secondary = TRUE)

# Suppress values below 10
suppresssed_data2 = sample_data |>
  primary_suppression(suppression_variable = number,
                      no_secondary = TRUE,
                      min_unsuppressed = 10)

# Drop the original, unsuppressed column
suppresssed_data3 = sample_data |>
  primary_suppression(suppression_variable = number,
                      no_secondary = TRUE,
                      keep_original = FALSE)

# Apply both primary and secondary suppression

sec_suppressed = sample_data |>
  primary_suppression(suppression_variable = number,
                      no_secondary = FALSE) |>
  # We want to apply secondary suppression within SIMD
  # Therefore group by all categories EXCEPT SIMD
  secondary_suppression(suppression_variable = number,
                        health_board, date)

}
