% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_axis_lists.R
\name{x_categorical}
\alias{x_categorical}
\title{Produces a list for a categorical x axis layout in a plotly chart}
\usage{
x_categorical()
}
\value{
A list
}
\description{
This function is internally used by \code{phs_plotly_layout()}
}
\keyword{Internal}
