% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_axis_lists.R
\name{y_proportion}
\alias{y_proportion}
\title{Produces a list for a proportion y axis layout in a plotly chart}
\usage{
y_proportion()
}
\value{
A list
}
\description{
This function is internally used by \code{phs_plotly_layout()}
}
\keyword{Internal}
