% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_linechart.R, R/single_linechart.R
\name{comparison_linechart}
\alias{comparison_linechart}
\title{Create a comparison line chart}
\usage{
comparison_linechart(
  data,
  x_axis,
  y_axis,
  category_var,
  palette = c("#3F3685", "#9B4393", "#1E7F84", "#83BB26", "#3393DD", "#6B5C85")
)

comparison_linechart(
  data,
  x_axis,
  y_axis,
  category_var,
  palette = c("#3F3685", "#9B4393", "#1E7F84", "#83BB26", "#3393DD", "#6B5C85")
)
}
\arguments{
\item{data}{The dataframe you want to create a chart from}

\item{x_axis}{The x axis variable (usually a date)}

\item{y_axis}{A y axis variable (usually a number)}

\item{category_var}{The category variable you want to split by, like color in
plotly or fill in ggplot2}

\item{palette}{A vector of hex codes for changing the palette. Default is PHS
colours}
}
\value{
A plotly chart

A plotly chart
}
\description{
Creates a generic comparison chart in plotly. This function makes it quick
and easy to reuse similar code, such as in a dashboard

Creates a generic single line chart in plotly. This function makes it quick
and easy to reuse similar code, such as in a dashboard
}
\examples{
## Set-up an example chart
library(dplyr)
library(phsshinyextras)

data("sample_data")
# Using the in-built sample data set to compare 3 Health Boards
example_data = filter(sample_data,
                      health_board \%in\% c("NHS Lothian",
                                          "NHS Ayrshire and Arran",
                                          "NHS Orkney"),
                      simd == "Total")
# Standard comparison line chart
example_data |>
  comparison_linechart(x_axis = date,
                       y_axis = number,
                       category_var = health_board)


# Creating a tooltip beforehand as a variable
tooltip_example = example_data |>
  mutate(tooltip = paste0(
    "Date: ", date, "<br>",
    scales::percent(proportion, accuracy = 1), " of records were in ",
    health_board
  ))

# Automatically picks up the tooltip by its name
tooltip_example |>
  comparison_linechart(x_axis = date,
                       y_axis = number,
                       category_var = health_board,
                       # Updating the palette
                       palette = c("Red", "Green", "Blue"))
## Set-up an example chart
library(dplyr)
library(phsshinyextras)

data("sample_data")
# Using the in-built sample data set to compare 3 Health Boards
example_data = filter(sample_data,
                      health_board \%in\% c("NHS Lothian",
                                          "NHS Ayrshire and Arran",
                                          "NHS Orkney"),
                      simd == "Total")
# Standard comparison line chart
example_data |>
  comparison_linechart(x_axis = date,
                       y_axis = number,
                       category_var = health_board)


# Creating a tooltip beforehand as a variable
tooltip_example = example_data |>
  mutate(tooltip = paste0(
    "Date: ", date, "<br>",
    scales::percent(proportion, accuracy = 1), " of records were in ",
    health_board
  ))

# Automatically picks up the tooltip by its name
tooltip_example |>
  comparison_linechart(x_axis = date,
                       y_axis = number,
                       category_var = health_board,
                       # Updating the palette
                       palette = c("Red", "Green", "Blue"))
}
