% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_suppression.R
\name{primary_suppression}
\alias{primary_suppression}
\title{Apply primary suppression to numbers > 0 and < 5}
\usage{
primary_suppression(data, suppression_variable, no_secondary = FALSE)
}
\arguments{
\item{data}{A dataset to suppress}

\item{suppression_variable}{A numeric/integer column to suppress}

\item{no_secondary}{TRUE means you don't need to apply secondary suppression}
}
\description{
Apply primary suppression to numbers > 0 and < 5
}
\examples{
man/examples/primary_suppression.R

}
